Isolate download logic from service lifecycle
