+++ app/src/main/java/com/topjohnwu/magisk/view/Notifications.kt
@@ -11,7 +11,7 @@ import androidx.core.content.getSystemService
 import androidx.core.graphics.drawable.toIcon
 import com.topjohnwu.magisk.R
 import com.topjohnwu.magisk.core.di.AppContext
-import com.topjohnwu.magisk.core.download.DownloadService
+import com.topjohnwu.magisk.core.download.DownloadManager
 import com.topjohnwu.magisk.core.download.Subject
 import com.topjohnwu.magisk.core.ktx.getBitmap
 import com.topjohnwu.magisk.core.ktx.selfLaunchIntent
@@ -67,7 +67,7 @@ object Notifications {
 
     fun updateAvailable() {
         AppContext.apply {
-            val intent = DownloadService.getPendingIntent(this, Subject.App())
+            val intent = DownloadManager.getPendingIntent(this, Subject.App())
             val bitmap = getBitmap(R.drawable.ic_magisk_outline)
             val builder = if (SDK_INT >= Build.VERSION_CODES.O) {
                 Notification.Builder(this, UPDATE_CHANNEL)
